+++ ASDKSample/ASDKSample/RootViewController.swift
@@ -73,9 +73,6 @@ class RootViewController: UITableViewController {
     private var dataSource: [Product] = []
     private var onScannerResult: ((_ number: String?, _ date: String?) -> Void)?
 
-    // State
-    private weak var buyProductsVieController: BuyProductsViewController?
-
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -190,12 +187,21 @@ class RootViewController: UITableViewController {
             cardListViewConfigration.alertViewHelper = self
         }
 
-        if let sdk = try? SdkAssembly.assembleUIsdk(creds: AppSetting.shared.activeSdkCredentials) {
-            // открыть экран сиска карт
-            addCardListView(sdk, AppSetting.shared.activeSdkCredentials.customerKey, cardListViewConfigration)
-            // или открыть экран добавлени карты
-            // addCardView(sdk, customerKey, cardListViewConfigration)
+        if let sdk = try? SdkAssembly.assembleUISDK(credential: AppSetting.shared.activeSdkCredentials) {
+            sdk.addCardNeedSetCheckTypeHandler = {
+                AppSetting.shared.addCardChekType
+            }
 
+            sdk.presentCardList(
+                on: self,
+                customerKey: AppSetting.shared.activeSdkCredentials.customerKey,
+                configuration: cardListViewConfigration
+            )
+        }
+    }
+
+    @IBAction func openAddCard(_ sender: UIBarButtonItem) {
+        if let sdk = try? SdkAssembly.assembleUISDK(credential: AppSetting.shared.activeSdkCredentials) {
             sdk.addCardNeedSetCheckTypeHandler = {
                 AppSetting.shared.addCardChekType
             }
@@ -227,7 +233,7 @@ extension RootViewController: AcquiringAlertViewProtocol {
 private extension RootViewController {
 
     private func showSpbQrCollector() {
-        if let sdk = try? SdkAssembly.assembleUIsdk(creds: AppSetting.shared.activeSdkCredentials) {
+        if let sdk = try? SdkAssembly.assembleUISDK(credential: AppSetting.shared.activeSdkCredentials) {
             let viewConfigration = AcquiringViewConfiguration()
             viewConfigration.viewTitle = Loc.Title.qrcode
 
@@ -236,22 +242,29 @@ private extension RootViewController {
     }
 
     private func showBuyProductsViewController(rowIndex: Int) {
-        if let sdk = try? SdkAssembly.assembleUIsdk(creds: AppSetting.shared.activeSdkCredentials) {
-            let product = dataSource[rowIndex]
+        let credential = AppSetting.shared.activeSdkCredentials
 
-            let storyboard = UIStoryboard(name: "Main", bundle: .main)
-            guard let viewController = storyboard.instantiateViewController(
-                withIdentifier: String(describing: BuyProductsViewController.self)
-            ) as? BuyProductsViewController
-            else {
-                return
-            }
-
-            viewController.scaner = self
-            viewController.sdk = sdk
-            viewController.customerKey = AppSetting.shared.activeSdkCredentials.customerKey
-            viewController.products = [product]
-            navigationController?.pushViewController(viewController, animated: true)
+        guard let coreSDK = try? SdkAssembly.assembleCoreSDK(credential: credential),
+              let uiSDK = try? SdkAssembly.assembleUISDK(credential: credential) else {
+            fatalError("Could not assemble SDK")
         }
+
+        let product = dataSource[rowIndex]
+
+        let storyboard = UIStoryboard(name: "Main", bundle: .main)
+
+        guard let viewController = storyboard.instantiateViewController(
+            withIdentifier: String(describing: BuyProductsViewController.self)
+        ) as? BuyProductsViewController
+        else {
+            fatalError("Could not instantiate BuyProductsViewController")
+        }
+
+        viewController.scaner = self
+        viewController.coreSDK = coreSDK
+        viewController.uiSDK = uiSDK
+        viewController.customerKey = credential.customerKey
+        viewController.products = [product]
+        navigationController?.pushViewController(viewController, animated: true)
     }
 }