+++ TinkoffASDKUI/TinkoffASDKUI/PaymentModule/Payments/CardPaymentProcess.swift
@@ -34,15 +34,6 @@ final class CardPaymentProcess: PaymentProcess {
 
     private weak var delegate: PaymentProcessDelegate?
 
-    private var customerEmail: String? {
-        switch paymentFlow {
-        case let .full(paymentOptions):
-            return paymentOptions.customerOptions?.email
-        case let .finish(_, customerOptions):
-            return customerOptions?.email
-        }
-    }
-
     init(
         paymentsService: IAcquiringPaymentsService,
         threeDsService: IAcquiringThreeDSService,
@@ -137,7 +128,7 @@ private extension CardPaymentProcess {
             let data = FinishAuthorizeData(
                 paymentId: payload.paymentId,
                 paymentSource: paymentSource,
-                infoEmail: customerEmail
+                infoEmail: paymentFlow.customerOptions?.email
             )
             finishAuthorize(data: data)
         case .parentPayment, .yandexPay:
@@ -161,7 +152,7 @@ private extension CardPaymentProcess {
                 let data = FinishAuthorizeData(
                     paymentId: paymentId,
                     paymentSource: self.paymentSource,
-                    infoEmail: self.customerEmail,
+                    infoEmail: self.paymentFlow.customerOptions?.email,
                     deviceInfo: deviceInfo,
                     threeDSVersion: payload.version
                 )
@@ -171,7 +162,7 @@ private extension CardPaymentProcess {
             let data = FinishAuthorizeData(
                 paymentId: paymentId,
                 paymentSource: paymentSource,
-                infoEmail: customerEmail
+                infoEmail: paymentFlow.customerOptions?.email
             )
             finishAuthorize(data: data, threeDSVersion: payload.version)
         }