+++ TinkoffASDKUI/TinkoffASDKUI/SDK/AcquiringUISDK.swift
@@ -238,7 +238,7 @@ public class AcquiringUISDK: NSObject {
 
         yandexPayButtonContainerFactoryProvider = YandexPayButtonContainerFactoryProvider(
             flowAssembly: YandexPayPaymentFlowAssembly(
-                yandexPayActivityAssebmly: YandexPayPaymentSheetAssembly(
+                yandexPayPaymentSheetAssembly: YandexPayPaymentSheetAssembly(
                     paymentControllerAssembly: paymentControllerAssembly
                 )
             ),